error is good now

+++ src/src/main/java/bgu/spl/mics/application/objects/TrackedObject.java
@@ -9,11 +9,11 @@ import java.util.List;
  */
 public class TrackedObject {
     private final String id;
-    private final long time;
+    private final int time;
     private final String description;
     List<CloudPoint> cloudPoints;
 
-    public TrackedObject(String id, long time, String description, List<CloudPoint> cloudPoints) {
+    public TrackedObject(String id, int time, String description, List<CloudPoint> cloudPoints) {
         this.id = id;
         this.time = time;
         this.description = description;
@@ -24,7 +24,7 @@ public class TrackedObject {
         return id;
     }
 
-    public long getTime() {
+    public int getTime() {
         return time;
     }
